End Of Section 5 - IdentityServer

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -7,6 +7,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,13 +52,14 @@ public class AuctionsController :ControllerBase
 
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
         //TODO : Add current user as a seller
 
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -74,6 +76,7 @@ public class AuctionsController :ControllerBase
             new {auction.Id}, newAuction);
     }
 
+    [Authorize]
     [HttpPut ("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -82,7 +85,7 @@ public class AuctionsController :ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model?? auction.Item.Model;
@@ -99,6 +102,7 @@ public class AuctionsController :ControllerBase
 
     }
 
+    [Authorize] 
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -106,7 +110,7 @@ public class AuctionsController :ControllerBase
 
         if (auction == null) return NotFound();
 
-        //TOD: Check Seller Unsername
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
